refactor(dashboard): move shell constants to neutral module to clean RSC boundary
CR detectó que `DashboardSidebarNav` (client) importaba
`navItemClass`, `navItemActiveClass` y `type DashboardRole` desde
`DashboardShell` (server async). En Next 15+ esto puede causar
problemas: un módulo client no debería importar de un módulo que
podría ser server-only (aunque acá `DashboardShell` no lo marca
explícitamente, la convención es importar constants client-safe desde
módulos neutros).

Fix: crear `src/components/dashboard/shell-constants.ts` sin
`"use client"` ni `"server-only"`. Mueve las 3 exports compartidas.
`DashboardShell` (server) y `DashboardSidebarNav` (client) importan
desde ahí. La re-exportación del tipo `DashboardRole` desde
`DashboardShell` se mantiene para no romper consumers externos
(`dashboard/layout.tsx` lo usaba).
